Allow forward references in constant statements
The constant statement now uses lax evaluation, but will only store a
constant if its value is known and does not contain any forward
references.

A set of all constant names, including those with unknown values, is
kept to prevent the user from redefining unknown constants. The set is
also used to provide a friendlier error message when an unknown constant
is evaluated in strict evaluation mode.